Refactor to use shared class.
